requirements.txt

PySide6==6.7.2
python-dotenv==1.0.1
alpaca-py==0.42.1
APScheduler==3.10.4
openai==1.109.0
yfinance==0.2.66
feedparser==6.0.12
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.4.2
joblib==1.3.2
pywin32==306
pytz==2024.1
